@@ -33,7 +33,7 @@
   },
   "config": {
     "platform": {
-      "php": "7.4.33"
+      "php": "8.1.32"
     },
     "allow-plugins": {
       "php-http/discovery": false
