@@ -25,6 +25,7 @@
   +--------------------------------------------------------------------+
  */
 
+use Civi\Api4\StateProvince;
 use Geocoder\Query\GeocodeQuery;
 use Geocoder\Model\AddressCollection;
 use CRM_Geocoder_ExtensionUtil as E;
@@ -266,6 +267,11 @@ class CRM_Utils_Geocode_Geocoder {
       'return' => array_keys($missingFields)
     ]);
     $inputValues = array_merge($existingAddress, $inputValues);
+
+    // clear an orphaned county
+    if (empty($inputValues['county']) || empty($inputValues['state_province'])) {
+      unset($inputValues['county'], $inputValues['county_id']);
+    }
   }
 
   /**
@@ -393,24 +399,88 @@ class CRM_Utils_Geocode_Geocoder {
           // not possible to determine state without the country.
           return '';
         }
-        $state = $firstResult->getAdminLevels()->get(1)->getCode();
-        if (!isset(\Civi::$statics[__CLASS__]['country_id'][$state])) {
-          try {
-            // Build our own static array as the core pseudoconstant does country limiting in a weird way.
-            \Civi::$statics[__CLASS__]['country_id'][$state] = civicrm_api3('StateProvince', 'getvalue', [
-              'return' => 'id',
-              'abbreviation' => $state,
-              'country_id' => $values['country_id'],
-            ]);
-          }
-          catch (CRM_Core_Exception $e) {
-            // We just won't worry about the state.
+        // Some providers return the code (abbreviation), others (eg nominatim/osm) return the name
+        $stateAbbreviation = $firstResult->getAdminLevels()->get(1)->getCode();
+        if (empty($stateAbbreviation) && !empty($firstResult->getAdminLevels()->get(1)->getName())) {
+          // Try to get Abbreviation from State Name.
+          $stateName = $firstResult->getAdminLevels()->get(1)->getName();
+          $stateProvince = StateProvince::get(FALSE)
+            ->addSelect('abbreviation', 'id')
+            ->addWhere('name', '=', $stateName)
+            ->addWhere('country_id', '=', $values['country_id'])
+            ->addWhere('is_active', '=', TRUE)
+            ->execute()
+            ->first();
+          if (empty($stateProvince['abbreviation'])) {
+            // We need the abbreviation to get the ID.
             return '';
           }
+          $stateAbbreviation = $stateProvince['abbreviation'];
+          $stateProvinceID = $stateProvince['id'];
         }
-        return \Civi::$statics[__CLASS__]['country_id'][$state];
+        // Now we have the state Abbreviation we can check the cache (and fill it if required)
+        if (!isset(\Civi::$statics[__CLASS__]['country_id'][$stateAbbreviation])) {
+          if (empty($stateProvinceID)) {
+            // We didn't already look up the ID, get it now
+            $stateProvince = StateProvince::get(FALSE)
+              ->addSelect('id')
+              ->addWhere('abbreviation', '=', $stateAbbreviation)
+              ->addWhere('country_id', '=', $values['country_id'])
+              ->addWhere('is_active', '=', TRUE)
+              ->execute()
+              ->first();
+            if (empty($stateProvince['id'])) {
+              // Not found. Abbreviation doesn't exist for this Country.
+              return '';
+            }
+            $stateProvinceID = $stateProvince['id'];
+          }
+          // Build our own static array as the core pseudoconstant does country limiting in a weird way.
+          \Civi::$statics[__CLASS__]['country_id'][$stateAbbreviation] = $stateProvinceID;
+        }
+        return \Civi::$statics[__CLASS__]['country_id'][$stateAbbreviation];
 
+      case 'county_id':
+        $stateAbbreviation = self::getValueFromResult('state_province_id', $result, $values);
+        $county = self::getAdminLevelByType($firstResult, 'county', 2);
+
+        if ($stateAbbreviation && $county) {
+          $id = CRM_Core_DAO::singleValueQuery('SELECT id FROM civicrm_county WHERE state_province_id = %1 AND name = %2 AND is_active=1', [
+            1 => [$stateAbbreviation, 'Integer'],
+            2 => [$county, 'String']
+          ]);
+          return $id ?: NULL;
+        }
+        return NULL;
     }
+  }
+
+  /**
+   * Get an admin level name by type.
+   * The name will end with the $type, e.g. "Lake County"...
+   * ...and the $type will be removed on return, e.g. "Lake".
+   * Can optionally return a name for a $defaultLevel if $type is not found in this way.
+   *
+   * @param Address $address Address to get admin level from.
+   * @param string $type The type of admin level to get, e.g. "County".
+   * @param integer $defaultLevel If $type is not found, use this level. 0 for no default level.
+   * @return mixed The requested value, or NULL.
+   */
+  protected static function getAdminLevelByType($address, $type, $defaultLevel = 0) {
+    $levels = $address->getAdminLevels();
+    $ew = strtolower(" $type");
+
+    foreach ($levels->all() as $level) {
+      $name = $level->getName();
+
+      if (str_ends_with(strtolower($name), $ew)) {
+        return substr($name, 0, -strlen($ew));
+      }
+    }
+    if ($defaultLevel > 0 && $levels->has($defaultLevel)) {
+      return $levels->get($defaultLevel)->getName();
+    }
+    return NULL;
   }
 
 
@@ -600,7 +670,9 @@ class CRM_Utils_Geocode_Geocoder {
       try {
         $provider = self::getProviderClass($geocoder);
         $geocoderObj = new \Geocoder\StatefulGeocoder($provider, $locale);
-        $result = $geocoderObj->geocodeQuery(GeocodeQuery::create(urlencode($geocodableAddress)));
+        // Not do not url_encode as it breaks OpenStreetMap
+        // https://github.com/eileenmcnaughton/org.wikimedia.geocoder/issues/61
+        $result = $geocoderObj->geocodeQuery(GeocodeQuery::create($geocodableAddress));
         return [
           'geo_code_1' => $result->first()->getCoordinates()->getLatitude(),
           'geo_code_2' => $result->first()->getCoordinates()->getLongitude(),
