@@ -63,52 +63,26 @@ function geocoder_civicrm_geo_managed(&$entities) {
 }
 
 /**
- * Make sure all our geocoders are in managed.
+ * @param string $formName
+ * @param CRM_Admin_Form_Setting_Mapping $form
  *
- * Historically they weren't because the managed system didn't
- * support 'update' => 'never' and kept enabling / reverting them.
- * But now, for a few releases, we will hold a routine to help convert.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_managed
+ * @return void
  */
-function geocoder_civicrm_managed(&$entities) {
-  if (!empty(Civi::$statics['geocoder_civicrm_managed']['check'])) {
-    return;
-  }
-  $geocoders = CRM_Core_DAO::executeQuery('SELECT id, name FROM civicrm_geocoder')->fetchAll();
-  $ids = [];
-  foreach ($geocoders as $geocoder) {
-    $ids[$geocoder['name']] = $geocoder['id'];
-  }
-  if (!empty($ids)) {
-    $managedEntities = CRM_Core_DAO::executeQuery("SELECT name FROM civicrm_managed WHERE module = 'org.wikimedia.geocoder' AND entity_type = 'Geocoder' AND entity_id IN (" . implode(',', $ids) . ')')->fetchAll();
-    foreach ($managedEntities as $managedEntity) {
-      if (isset($ids[$managedEntity['name']])) {
-        unset($ids[$managedEntity['name']]);
-      }
-    }
-    foreach ($ids as $name => $id) {
-      CRM_Core_DAO::executeQuery("INSERT INTO civicrm_managed (module, name, entity_type, entity_id) VALUES('org.wikimedia.geocoder', '$name', 'Geocoder', $id)");
+function geocoder_civicrm_buildForm($formName, $form): void {
+  if ($formName === 'CRM_Admin_Form_Setting_Mapping') {
+    $form->setTitle(E::ts('Mapping provider'));
+    $warning = '<p>' . E::ts('You have the geocoder extension installed, which is not configured on this page. You need to <a href="%1">configure geocoding at the extension page</a> or disable the extension',
+      [1 => CRM_Utils_System::url('civicrm/admin/geocoders')]
+    ) . '</p>';
+    $warning .= '<p>' . E::ts('<a href="https://docs.civicrm.org/geocoder/en/latest/"> Read more') . '</p>';
+    $markup = '<div class="help">' . $warning . ' </div>';
+    $geocoder = $form->getElementValue('geoProvider');
+    if ($geocoder === ['Geocoder']) {
+      $form->getElement('geoProvider')->freeze();
+      $form->getElement('geoAPIKey')->freeze();
+      CRM_Core_Region::instance('page-header')->addMarkup($markup);
     }
   }
-  Civi::$statics['geocoder_civicrm_managed']['check'] = TRUE;
-
-}
-
-/**
- * Implements hook_civicrm_entityTypes.
- *
- * @link https://docs.civicrm.org/dev/en/latest/hooks/hook_civicrm_entityTypes
- *
- * @param array $entityTypes
- *   Registered entity types.
- */
-function geocoder_civicrm_entityTypes(&$entityTypes) {
-  $entityTypes['CRM_Geocoder_DAO_Geocoder'] = [
-    'name' => 'Geocoder',
-    'class' => 'CRM_Geocoder_DAO_Geocoder',
-    'table' => 'civicrm_geocoder',
-  ];
 }
 
 /**
